+++ src/tiled/editabletilelayer.cpp
@@ -93,16 +93,7 @@ int EditableTileLayer::flagsAt(int x, int y) const
 
 EditableTile *EditableTileLayer::tileAt(int x, int y) const
 {
-    if (Tile *tile = cellAt(x, y).tile()) {
-        auto tileset = tile->tileset()->sharedFromThis();
-
-        if (auto tilesetDocument = TilesetDocument::findDocumentForTileset(tileset)) {
-            EditableTileset *editable = tilesetDocument->editable();
-            return EditableManager::instance().editableTile(editable, tile);
-        }
-    }
-
-    return nullptr;
+    return EditableManager::instance().editableTile(cellAt(x, y).tile());
 }
 
 TileLayerEdit *EditableTileLayer::edit()