+++ NEWS.md
@@ -42,6 +42,7 @@
 * Scripting: Fixed possible crash after creating tilesets from script (#3229)
 * Scripting: Fixed possible crash in TileMap.autoMap
 * Scripting: Fixed dialog window titles to show on macOS (#2910)
+* Scripting: Fixed tileset or tile references for maps loaded from script
 * Fixed the logic for handling group layer parallax factors (with LilithSilver, #3125)
 * Fixed keyboard modifiers getting stuck for Terrain Brush (#2678)
 * Fixed debug messages showing in the Console and Issues views